Code refactor
* Removed deprecated code
* Removed excessive injections

* Moved all models under _models
* Moved all services under _services
