+++ src/app/_services/chat.service.ts
@@ -1,10 +1,9 @@
 import { Injectable, EventEmitter } from '@angular/core';
-import { Chat } from "./chat";
-import { CHATS } from "./mock-chats";
+import { Chat } from "../_models/chat";
 import { Observable, of } from 'rxjs';
-import { HttpClient, HttpHeaders } from '@angular/common/http';
+import { HttpHeaders } from '@angular/common/http';
 import { AngularFirestore, AngularFirestoreCollection, AngularFirestoreDocument } from '@angular/fire/firestore';
-import { AuthService } from './_services/auth.service';
+import { AuthService } from './auth.service';
 import { firestore } from 'firebase';
 
 @Injectable({
@@ -42,7 +41,6 @@ export class ChatService {
 
     }
     this.chatsCollection.doc(cid).set(data);
-    // return this.http.post<Chat>(this.chatUrl, chat, this.httpOptions);
   }
 
   getChatSettings(cid: string): Observable<Chat> {
@@ -95,7 +93,6 @@ export class ChatService {
   }
 
   constructor(
-    private http: HttpClient,
     private database: AngularFirestore,
     private auth: AuthService
   ) {
