--- src/app/mock-chats.ts
@@ -1,10 +1,0 @@
-// Import Chats data model
-import { Chat } from './chat';
-
-// Create fake array for mock chats called CHATS, use Chat data model array
-export const CHATS: Chat[] = [
-    // {id: 1, name: 'Cloudrunner', picture: 'https://png.pngtree.com/png-vector/20191108/ourlarge/pngtree-fast-leg-run-runner-running-blue-icon-on-abstract-cloud-back-png-image_1966561.jpg'},
-    // {id: 2, name: 'Stormers', picture: 'https://png.pngtree.com/png-vector/20190618/ourlarge/pngtree-lightning-icon-png-image_1506453.jpg'},
-    // {id: 3, name: 'SunTimes', picture: 'https://img.lovepik.com/original_origin_pic/18/04/08/4000c4cfe5096aebc5283e6d50a805ae.png_wh860.png'}
-
-]
