+++ src/app/navbar/navbar.component.html
@@ -2,8 +2,8 @@
     <a class="navbar-brand col-3 mr-0" href="#">MessageApp</a>
     <button class="navbar-toggler mr-1" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
         <span class="navbar-toggler-icon"></span>
-      </button>
-      <div class="collapse navbar-collapse" id="navbarSupportedContent">
+    </button>
+    <div class="collapse navbar-collapse" id="navbarSupportedContent">
         <ul class="navbar-nav px-3 ml-auto" *ngIf="auth.user$ | async as user; else publicNavbar">
             <small class="d-flex align-items-center">
                 <li class="nav-item text-nowrap">
@@ -24,5 +24,5 @@
             </li>
         </ul>
         </ng-template>
-      </div>
+    </div>
 </nav>
