+++ src/app/users/login/login.component.ts
@@ -39,5 +39,4 @@ export class LoginComponent implements OnInit {
       this.roles = this.tokenStorageService.getUser().roles;
     }
   }
-
 }
