+++ src/environments/environment.prod.ts
@@ -1,11 +1,5 @@
 export const environment = {
-  production: true,
-  firebase: {
-    apiKey: "AIzaSyDBvnLWYWPf9dO20JkCbclfZhrOptE4w7g",
-    authDomain: "message-app-ff56f.firebaseapp.com",
-    databaseURL: "https://message-app-ff56f.firebaseio.com",
-    projectId: "message-app-ff56f",
-    storageBucket: "message-app-ff56f.appspot.com",
-    messagingSenderId: "338385062636"
-  }
+  // API key was leaked to public, changed key
+  production: true
+
 };
